changed some variables since the csv data for motor units were in bytes and not float. Now they should have decimals. I have fixed scaling of i2c data. so it is multiplied when sent, and divided when received

+++ CCU/src/i2c_master.h
@@ -21,8 +21,8 @@ class I2CMaster {
 public:
     I2CMaster();
     void begin(); 
-    bool sendParam(uint8_t slave_adress, uint8_t mode, uint8_t kp, uint8_t ki, uint8_t kd);
-    bool sendSetpoint(uint8_t slave_adress, uint8_t setpoint);
+    bool sendParam(uint8_t slave_adress, uint8_t mode, float kp, float ki, float kd);
+    bool sendSetpoint(uint8_t slave_adress, float setpoint);
     bool requestData(uint8_t slave_adress, MUData& data);
 
 private:
